fixed failure use cases treatment
There was a problem with saving failure errors to results file.
The reason was that after 1st failure _.map continued to next flavor
and when it failed result was written to same file simultaneously. that
caused the result file to be empty from time to time.
Purpose of changes: write result file once. skip rest of flavours after
1st failure (when getting chunklist/segment from packager).
As part of the fix, long methods were split.

work not finished.