support reset recordedEntryId if flavor set changed between sessions
Covered use cases of recording start when old session content exists:
1. previous session used flavor set according to transcoding profile
but this one uses different set. RecordedEntryId is reset and a new
session starts.
2. previous session used flavor set different than transcoding profile
but this one uses aligned flavors set. RecordedEntryId is reset and a
new session starts.
3. current and previous sessions uses different flavors set than
transcoding profile and the flavors sets are the same (recording starts
as if flavors are aligned). No change from regular flow.
4. current and previous sessions uses different flavors set than
transcoding profile and the flavors sets are not the same.
RecordedEntryId is reset and a new session starts.
5. previous and current sessions have flavor set aligned with
transcoding profile. This is the common use case. No change in the flow.
this commit is before testing.