merge to 1.17 PLAT-7109 recording lost when flavors has changed between sessions (#509)
* support reset recordedEntryId if flavor set changed between sessions

Covered use cases of recording start when old session content exists:
1. previous session used flavor set according to transcoding profile
but this one uses different set. RecordedEntryId is reset and a new
session starts.
2. previous session used flavor set different than transcoding profile
but this one uses aligned flavors set. RecordedEntryId is reset and a
new session starts.
3. current and previous sessions uses different flavors set than
transcoding profile and the flavors sets are the same (recording starts
as if flavors are aligned). No change from regular flow.
4. current and previous sessions uses different flavors set than
transcoding profile and the flavors sets are not the same.
RecordedEntryId is reset and a new session starts.
5. previous and current sessions have flavor set aligned with
transcoding profile. This is the common use case. No change in the flow.
this commit is before testing.

* typo fix

* typo fix

* added missing 'this'

* more changes in RecordingManager and name refactor

*  code update during testing

* more changes based on testing

* returned splitRecordedEntry to RecordingEntrySession

+ changes in verifyStreamMatchPreviousFlavorSet() to correct compare
previous and current flavors list (by name)

* pass recordingEntrySession to calls to start recording

* handle case that recording didn't start and stopRecording is called

the use cases that stopRecording can be called before recording started:
1) due to transcoding profile change
2) live provisioning different than previous live stream (flavor set
changed but not transcoding profile)
3) fast stop start of live without recording

* updated stopRecording to defer recording that didn't start from very short duration

* Changes after code review + making starRecording small and readable

tests in progress. 24-7 flow verified already.

* add missing mp4FilesQueue to call to getRecordingEntrySession()

* another fix to properly propagate mp4FilesQueue while split recording done

* when live provision changed don't use queue of previous recording session

removed passing mp4FilesQueue.lock from cal to getRecordingEntrySession

* added support to use case that flavors were added between to consecutive recordings

* fix handling use cases of very short stop start with change in flavor set

Before the change in any use case that recordedEntrySession was found
in recordingList null was returned and the flow to verify new flavor
set matches the recording set in use, was not checked.

* fix flow to avoid restarting recording for already started session

+ moved verifyStreamMatchPreviousFlavorSet to RecordingEntrySession

* added log message when live started and recording session exists

* split lastTimestampSessionModification

replaced by isTimeToStopSession() and isTimeToCreateNewSession()

* added space to RecordingEntrySession logger header

* changes based on Gad Y.'s review

after testing.

* updated the way last recording flavors are retreived

Instead of reading the recording directory tree the playlist.json is
read and sequences ids retreived.
Didn’t remove

* updates after Guy J's review

* code review updates

after testing

* changes after code review with Guy and some tests

main change: get previous recording flavors from playlist.json using
PlaylistGenerator class

* updates based on latest tests

main change: check if recording content path exist before calling
handleFlavorsMismatch()

* fixed handling new session and check flavor set using playlistGenerator

PlaylistGenerator swallow promise.reject() if recording playlist.json
is not found. This is the opposite of required behavior for flow of
verify flavor set mismatch. To resolve that, I added check if recording
path exists before checking if there is flavor set mismatch.

* if recorded entry id mismatch is found in getRecordedEntrySession() return rejected promise

Recording start flow was updated in a way that it is unexpected to find
existing RecordedEntrySession with different RecordedEntryId than in
EntryObj , in a call to getRecordedEntrySession().

* plat-7944

delete recording session object if flavor mismatch found and session
not started yet

* fix python command in liveRecorder script